requirements.txt
aiohappyeyeballs>=2.6.1
aiohttp>=3.13.2
aiosignal>=1.4.0
annotated-types>=0.7.0
anyio>=4.11.0
arrow>=1.4.0
asgiref>=3.8.1
asttokens>=3.0.0
attrs>=25.3.0
backoff>=2.2.1
bcrypt>=5.0.0
beautifulsoup4>=4.14.2
bitarray>=3.4.0
blinker>=1.9.0
bs4>=0.0.1
build>=1.3.0
cachetools>=6.2.1
canvasapi>=3.3.0
certifi>=2025.8.3
cffi>=2.0.0
charset-normalizer>=3.4.2
chromadb>=1.3.4
ckzg>=2.1.1
click>=8.3.0
colorama>=0.4.6
coloredlogs>=15.0.1
comm>=0.2.2
cytoolz>=1.0.1
dataclasses-json>=0.6.7
debugpy>=1.8.12
decorator>=5.2.1
distro>=1.9.0
dj-database-url>=3.0.1
Django>=5.1.5
django-cors-headers>=4.6.0
django-environ>=0.12.0
djangorestframework>=3.15.2
dnspython>=2.8.0
docstring_parser>=0.17.0
durationpy>=0.10
email-validator>=2.3.0
eth-account>=0.13.7
eth-hash>=0.7.1
eth-keyfile>=0.8.1
eth-keys>=0.7.0
eth-rlp>=2.2.0
eth-typing>=5.2.1
eth-utils>=5.3.0
eth_abi>=5.2.0
executing>=2.2.0
fastapi>=0.111.0
fastapi-cli>=0.0.13
filelock>=3.20.0
Flask>=3.1.2
flask-cors>=6.0.1
flatbuffers>=25.9.23
frozenlist>=1.8.0
fsspec>=2025.10.0
generativeai>=0.0.1
google-ai-generativelanguage>=0.6.15
google-api-core>=2.28.1
google-api-python-client>=2.187.0
google-auth>=2.43.0
google-auth-httplib2>=0.2.1
google-cloud-aiplatform>=1.71.1
google-cloud-bigquery>=3.38.0
google-cloud-core>=2.5.0
google-cloud-firestore>=2.21.0
google-cloud-resource-manager>=1.15.0
google-cloud-storage>=2.19.0
google-crc32c>=1.7.1
google-genai>=1.49.0
google-generativeai>=0.8.5
google-resumable-media>=2.7.2
googleapis-common-protos>=1.72.0
greenlet>=3.2.4
grpc-google-iam-v1>=0.14.3
grpcio>=1.76.0
grpcio-status>=1.62.3
gunicorn>=23.0.0
h11>=0.16.0
hexbytes>=1.3.0
hf-xet>=1.2.0
httpcore>=1.0.9
httplib2>=0.31.0
httptools>=0.6.4
httpx>=0.28.1
httpx-sse>=0.4.3
huggingface_hub>=1.1.2
humanfriendly>=10.0
idna>=3.10
importlib_metadata>=8.7.0
importlib_resources>=6.5.2
iniconfig>=2.1.0
ipykernel>=6.29.5
ipython>=8.32.0
itsdangerous>=2.2.0
jedi>=0.19.2
Jinja2>=3.1.6
jiter>=0.11.0
joblib>=1.4.2
jsonpatch>=1.33
jsonpointer>=3.0.0
jsonschema>=4.25.1
jsonschema-specifications>=2025.9.1
jupyter_client>=8.6.3
jupyter_core>=5.7.2
langchain>=1.0.4
langchain-classic>=1.0.0
langchain-community>=0.4.1
langchain-core>=1.0.3
langchain-text-splitters>=1.0.0
langgraph>=1.0.2
langgraph-checkpoint>=3.0.1
langgraph-prebuilt>=1.0.2
langgraph-sdk>=0.2.9
langsmith>=0.4.41
lxml>=6.0.2
markdown-it-py>=4.0.0
MarkupSafe>=3.0.3
marshmallow>=3.26.1
matplotlib-inline>=0.1.7
mdurl>=0.1.2
mmh3>=5.2.0
mpmath>=1.3.0
multidict>=6.7.0
mypy_extensions>=1.1.0
nest-asyncio>=1.6.0
networkx>=3.5
numpy>=2.2.4
oauthlib>=3.3.1
onnxruntime>=1.23.2
openai>=2.0.0
opentelemetry-api>=1.38.0
opentelemetry-exporter-otlp-proto-common>=1.38.0
opentelemetry-exporter-otlp-proto-grpc>=1.38.0
opentelemetry-proto>=1.38.0
opentelemetry-sdk>=1.38.0
opentelemetry-semantic-conventions>=0.59b0
orjson>=3.11.3
ormsgpack>=1.12.0
outcome>=1.3.0.post0
overrides>=7.7.0
packaging>=24.2
pandas>=2.2.3
parsimonious>=0.10.0
parso>=0.8.4
pillow>=12.0.0
platformdirs>=4.3.6
pluggy>=1.5.0
poly_eip712_structs>=0.0.1
posthog>=5.4.0
prompt_toolkit>=3.0.50
propcache>=0.4.1
proto-plus>=1.26.1
protobuf>=5.27.0
psutil>=7.0.0
psycopg2-binary>=2.9.10
pure_eval>=0.2.3
py_clob_client>=0.20.0
py_order_utils>=0.3.2
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pybase64>=1.4.2
pycparser>=2.23
pycryptodome>=3.22.0
pydantic>=2.12.4
pydantic-settings>=2.11.0
pydantic_core>=2.41.5
Pygments>=2.19.1
pyparsing>=3.2.5
PyPika>=0.48.9
pyproject_hooks>=1.2.0
pyreadline3>=3.5.4
PySocks>=1.7.1
pytest>=7.4.3
pytest-mock>=3.14.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pytz>=2025.2
PyYAML>=6.0.3
pyzmq>=26.2.1
referencing>=0.37.0
regex>=2024.11.6
requests>=2.32.5
requests-oauthlib>=2.0.0
requests-toolbelt>=1.0.0
rich>=14.1.0
rich-toolkit>=0.15.1
rlp>=4.1.0
rpds-py>=0.28.0
rsa>=4.9.1
scikit-learn>=1.6.1
scipy>=1.15.2
shapely>=2.1.2
shellingham>=1.5.4
six>=1.17.0
sniffio>=1.3.1
sortedcontainers>=2.4.0
soupsieve>=2.8
SQLAlchemy>=2.0.44
sqlparse>=0.5.3
stack-data>=0.6.3
starlette>=0.37.2
sympy>=1.14.0
tenacity>=9.1.2
threadpoolctl>=3.6.0
tokenizers>=0.22.1
toolz>=1.0.0
tornado>=6.4.2
tqdm>=4.67.1
traitlets>=5.14.3
trio>=0.30.0
trio-websocket>=0.12.2
typer>=0.19.2
typer-slim>=0.20.0
typing-inspect>=0.9.0
typing-inspection>=0.4.2
typing_extensions>=4.14.1
tzdata>=2025.1
ujson>=5.11.0
uritemplate>=4.2.0
urllib3>=2.5.0
uvicorn>=0.30.1
vertexai>=1.71.1
watchfiles>=1.1.0
wcwidth>=0.2.13
websocket-client>=1.8.0
websockets>=15.0.1
Werkzeug>=3.1.3
wfastcgi>=3.0.0
wsproto>=1.2.0
xxhash>=3.6.0
yarl>=1.22.0
zipp>=3.23.0
zstandard>=0.25.0
jinja2>=3.1.3
PyYAML>=6.0.1 